+++ lib/archive_validator/posts_validator.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class ArchiveValidator
+  class PostsValidator < CollectionValidator
+    def collection
+      posts
+    end
+
+    def entity_validator
+      PostValidator
+    end
+  end
+end