+++ Changelog.md
@@ -11,6 +11,7 @@
 ## Features
 * Add client-side cropping of profile image uploads [#7581](https://github.com/diaspora/diaspora/pull/7581)
 * Add client-site rescaling of post images if they exceed the maximum possible size [#7734](https://github.com/diaspora/diaspora/pull/7734)
+* Add backend for archive import [#7660](https://github.com/diaspora/diaspora/pull/7660)
 
 # 0.7.11.0
 
@@ -19,6 +20,7 @@
 ## Bug fixes
 
 ## Features
+* Add a manifest.json file as a first step to make diaspora* a Progressive Web App [#7998](https://github.com/diaspora/diaspora/pull/7998)
 
 # 0.7.10.0